requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Werkzeug==2.2.3
matplotlib==3.7.1
seaborn==0.12.2
pandas==2.0.3
reportlab==3.6.12
python-pptx==0.6.21
python-dotenv==1.0.1